Basic GTK window
