+++ hooya-gtk/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hooya"
+name = "hooya-gtk"
 version = "0.1.0"
 edition = "2021"
 authors = [ "wesl-ee <w@wesl.ee>" ]
@@ -7,13 +7,14 @@ authors = [ "wesl-ee <w@wesl.ee>" ]
 [dependencies]
 tonic = { version = "0.9" }
 prost = { version = "0.11" }
-cid = "0.10"
-ring = "0.16"
 tokio = { version = "1.0", features = [ "macros", "rt-multi-thread"] }
-sqlx = { version = "0.6", features = [ "runtime-tokio-native-tls" ] }
+tokio-stream = { version = "0.1" }
+hooya = { path = "../packages/hooya" }
+semver = "1.0"
+clap = { version = "4.3", features = [ "derive", "env", "cargo" ] }
 anyhow = "1.0"
-infer = "0.14"
 futures-util = "0.3"
+gtk = { version = "0.6.6", package = "gtk4", features = ["v4_8"] }
 
 [build-dependencies]
 tonic-build = "0.9"
